+++ tests/test_webservice.py
@@ -67,7 +67,7 @@ class TestWebservice(object):
                  'pdb_title': 'Catalase structure determined by electron crystallography of thin 3D crystals',
                  'atom_codes': 'PA,O1A,O2A,O5B,C5B,C4B,O4B,C3B,O3B,C2B,C1B,O3,PN,O1N,O2N,O5D,C5D,C4D,O4D,C3D,O3D,C2D,O2D,C1D,N1N,C2N,C3N,C7N,O7N,N7N,C4N,C5N,C6N',
                  'uniprot_acc': 'P00432',
-                'prot_chain': 'A', 'het_seq_nr': 602, 'het_code': 'NDP', 'prot_name': 'Catalase',
+                 'prot_chain': 'A', 'het_seq_nr': 602, 'het_code': 'NDP', 'prot_name': 'Catalase',
                  'ec_number': '1.11.1.6', 'frag_nr': 24, 'frag_id': '3j7u_NDP_frag24', 'rowid': 7059,
                  'uniprot_name': 'Catalase', 'nr_r_groups': 2, 'het_chain': 'A', 'hash_code': '6ef5a609fb192dba'}
             ]
@@ -90,6 +90,14 @@ class TestWebservice(object):
         eq_(self.app.app.config['matrix'], self.matrix)
         eq_(self.app.app.config['db_fn'], self.frags_db_fn)
         eq_(self.app.app.json_encoder, KripodbJSONEncoder)
+
+    def test_get_fragment_svg(self):
+        fragment_id = '3j7u_NDP_frag24'
+        with self.app.app.test_request_context():
+            result = server.get_fragment_svg(fragment_id, 400, 150)
+            ok_('NH' in result)
+            ok_('400' in result)
+            ok_('150' in result)
 
 
 class TestWebServiceClient(object):